Path entity and better project structure
