--- squall/background.py
@@ -1,1 +1,0 @@
-from starlette.background import BackgroundTasks as BackgroundTasks  # noqa
