--- squall/middleware/gzip.py
@@ -1,1 +1,0 @@
-from starlette.middleware.gzip import GZipMiddleware as GZipMiddleware  # noqa
