+++ tests/test_additional_response_extra.py
@@ -22,7 +22,7 @@ openapi_schema = {
     "openapi": "3.0.2",
     "info": {"title": "Squall", "version": "0.1.0"},
     "paths": {
-        "/items/": {
+        "/items": {
             "get": {
                 "responses": {
                     "200": {
@@ -31,7 +31,7 @@ openapi_schema = {
                     }
                 },
                 "summary": "Read Item",
-                "operationId": "read_item_items__get",
+                "operationId": "read_item_items_get",
             }
         }
     },
@@ -47,6 +47,6 @@ def test_openapi_schema():
 
 
 def test_path_operation():
-    response = client.get("/items/")
+    response = client.get("/items")
     assert response.status_code == 200, response.text
     assert response.json() == {"id": "foo"}
