+++ tests/test_exception_handlers.py
@@ -1,4 +1,4 @@
-from squall import HTTPException, Squall
+from squall import HTTPException, Num, Path, Squall
 from squall.exceptions import RequestPayloadValidationError
 from squall.responses import PrettyJSONResponse
 from squall.testclient import TestClient
@@ -27,8 +27,9 @@ def route_with_http_exception():
     raise HTTPException(status_code=400)
 
 
-@app.get("/request-validation/{param}/")
-def route_with_request_validation_exception(param: int):
+@app.get("/request-validation/{param}")
+def route_with_request_validation_exception(param: int = Path(valid=Num(ge=5))):
+    raise ValueError(param)
     pass  # pragma: no cover
 
 
@@ -39,14 +40,10 @@ def test_override_http_exception():
 
 
 def test_override_request_validation_exception():
-    response = client.get("/request-validation/invalid")
+    response = client.get("/request-validation/3")
     assert response.status_code == 400
     assert response.json() == {
         "details": [
-            {
-                "loc": ["path_params", "param"],
-                "msg": "Cast of `int` failed",
-                "val": "invalid",
-            }
+            {"loc": ["path_params", "param"], "msg": "Validation error", "val": 3}
         ]
     }
