+++ tests/test_infer_param_optionality.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 import pytest
 from squall import Squall
-from squall.router import Router
+from squall.routing.router import Router
 from squall.testclient import TestClient
 
 app = Squall()
