+++ tests/test_routing/test_sub_routing.py
@@ -1,11 +1,11 @@
-from squall.router import RootRouter, Router
+from squall.routing.router import RootRouter, Router
 
 
 def test_sub_router():
     v1_router = Router(prefix="/v1")
     user_router = Router(prefix="/user")
     user_router.add_api_route(
-        "/some/prefix(?P<some>[^/]+)/item/(?P<item>[^/]+)/(?P<more>[^/]+)-suffix",
+        "/some/{some}/item/{item}",
         methods=["GET", "POST"],
         endpoint=lambda *a: {},
     )
@@ -14,11 +14,6 @@ def test_sub_router():
     router = RootRouter(prefix="/api")
     router.include_router(v1_router)
 
-    assert router.routes[0].path == "".join(
-        [
-            "/api",
-            "/v1",
-            "/user",
-            "/some/prefix(?P<some>[^/]+)/item/(?P<item>[^/]+)/(?P<more>[^/]+)-suffix",
-        ]
+    assert router.routes[0].path.path == "".join(
+        ["/api", "/v1", "/user", "/some/{some}/item/{item}"]
     )
