+++ tests/test_ws_router.py
@@ -61,7 +61,7 @@ def test_router():
 
 def test_prefix_router():
     client = TestClient(app)
-    with client.websocket_connect("/prefix/") as websocket:
+    with client.websocket_connect("/prefix") as websocket:
         data = websocket.receive_text()
         assert data == "Hello, router with prefix!"
